requirements.txt
fastapi
uvicorn[standard]
requests
pywebview
pyserial
pynmea2